Fixed conversion 'error'

+++ src/arkode/arkode_lsrkstep_io.c
@@ -457,7 +457,7 @@ int lsrkStep_SetDefaults(ARKodeMem ark_mem)
   step_mem->stagemax       = 0;
   step_mem->ndomeigupdates = 0;
   step_mem->stagemaxlimit =
-    round(SUNRsqrt(ark_mem->reltol / (10.0 * ark_mem->uround)));
+    (int)round(SUNRsqrt(ark_mem->reltol / (10.0 * ark_mem->uround)));
   step_mem->stagemaxlimit =
     (step_mem->stagemaxlimit > 2) ? step_mem->stagemaxlimit : 2;
   step_mem->nstsig = 0;
